refactor: ProfileSnapShot domain 패키지 구조로 이동
