+++ src/main/java/com/example/bak/user/infra/persistence/command/ProfileJpaRepository.java
@@ -1,6 +1,6 @@
 package com.example.bak.user.infra.persistence.command;
 
-import com.example.bak.comment.application.command.port.dto.ProfileSnapShot;
+import com.example.bak.comment.domain.ProfileSnapShot;
 import com.example.bak.user.domain.Profile;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
